Refactor encode.rs (2,338 lines) into type-organized modules
Split the monolithic encode.rs into 7 focused submodules under encode/:
- block.rs: block, header, operational cert, VRF, protocol version encoding
- certificate.rs: certificates, credentials, relays, pool params
- governance.rs: voting procedures, proposal procedures, gov actions, DReps
- protocol_params.rs: protocol parameter update, cost models
- script.rs: native scripts, script refs, redeemers, witnesses, language views
- transaction.rs: transaction body, output, witness set, auxiliary data
- value.rs: value and multi-asset encoding

All public items re-exported from encode/mod.rs for full backwards compatibility.
No behavioral changes — pure refactoring.

Closes #29
